requirements.txt
apache-airflow==2.7.0
pandas==1.5.3
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
boto3==1.34.34
requests==2.31.0
